% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pisa_2006}
\alias{pisa_2006}
\title{Mathematics test scores of Spain, Germany and Estonia in the PISA 2006 test}
\format{A data frame with 25884 rows and 10 variables:
\describe{
  \item{year}{Year of the survey}
  \item{CNT}{Long country names}
  \item{STIDSTD}{Unique student id}
  \item{father_edu}{The father's highest achieved degree in the ISCED scale}
  \item{household_income}{The household's total income in categories}
  \item{avg_math}{The average math test score out of the 5 plausible values in Mathematics}
}}
\source{
A subset extracted from the \code{PISA2006lite} R package, https://github.com/pbiecek/PISA2012lite
}
\usage{
pisa_2006
}
\description{
A dataset containing the test scores and other household information of
students from Spain, Germany and Estonia from the PISA 2006 test.
}
\keyword{datasets}
